Cleanup label posthook
